chore: rename renderers to render
